Funcionando. Ou será que não?

+++ PromoLimit/Models/Produto.cs
@@ -7,7 +7,7 @@
         public bool Ativo { get; set; }
         public string Descricao { get; set; }
         public int Estoque { get; set; }
-        public int? Variacao { get; set; }
+        public long? Variacao { get; set; }
         public int Seller { get; set; }
     }
 }
